from unsigned int to unsigned long int for long numbers

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -8,7 +8,7 @@
  * Return: length of the number.
  */
 
-unsigned int Number_length(unsigned int n)
+unsigned int Number_length(unsigned long int n)
 {
 	int len = 0;
